* Fixed: prevent rest state lost at sit to chair. Include check mapid also in checking in inn position for rest state set/unset. Restore set rest state at enter to city (except underground cities) after client stop send zone updates at city enter (only send at leave). Implement this using check zone update by timer. * Move spell cast timer set to Spell::prepare to allow modify Spell::casttime for scripts, for examples, after spell object creating. * Fixed: use uint64 temporary variable instead uint32 for store spellFamilyMask. Old code not create problems currently but can later if not fix this. * Fixed: apply correctly by result (but maybe not by implememntation, but I not found another way at this moment) shaman talent that add percent to damage of weapon shaman enchantments in case physical damage per second enchantment.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@4445 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/Unit.cpp
@@ -5185,7 +5185,7 @@ bool Unit::IsDamageToThreatSpell(SpellEntry const * spellInfo) const
         return false;
 
     uint32 family = spellInfo->SpellFamilyName;
-    uint32 flags = spellInfo->SpellFamilyFlags;
+    uint64 flags = spellInfo->SpellFamilyFlags;
 
     if((family == 5 && flags == 256) ||                     //Searing Pain
         (family == 6 && flags == 8192) ||                   //Mind Blast
